Refactor JS
- Fix a JS error
- Show no password strength when no password
- Extract cron generator JS to module
- Extract tabs JS to module
