+++ package.json
@@ -27,8 +27,8 @@
 	},
 	"devDependencies": {
 		"@prettier/plugin-php": "^0.19.4",
-		"@typescript-eslint/eslint-plugin": "^5.59.0",
-		"@typescript-eslint/parser": "^5.59.0",
+		"@typescript-eslint/eslint-plugin": "^5.59.1",
+		"@typescript-eslint/parser": "^5.59.1",
 		"cssnano": "^6.0.0",
 		"esbuild": "^0.17.18",
 		"eslint": "^8.39.0",
@@ -50,7 +50,7 @@
 		"stylelint": "^15.6.0",
 		"stylelint-config-standard": "^33.0.0",
 		"typescript": "^5.0.4",
-		"vitepress": "1.0.0-alpha.73",
+		"vitepress": "1.0.0-alpha.74",
 		"vue": "^3.2.47"
 	},
 	"browserslist": [
