+++ web/js/pages/edit_mail.js
@@ -1,4 +1,4 @@
-document.getElementById('vstobjects').addEventListener('submit', function () {
+document.querySelector('#vstobjects').addEventListener('submit', () => {
 	$('input[disabled]').each(function (i, elm) {
 		var copy_elm = $(elm).clone(true);
 		$(copy_elm).attr('type', 'hidden');
