+++ web/js/pages/edit_web.js
@@ -261,7 +261,7 @@ $(function () {
 		}
 	});
 
-	document.getElementById('vstobjects').addEventListener('submit', function () {
+	document.querySelector('#vstobjects').addEventListener('submit', () => {
 		$('input[disabled]').each(function (i, elm) {
 			var copy_elm = $(elm).clone(true);
 			$(copy_elm).attr('type', 'hidden');
