+++ web/templates/pages/add_mail_acc.php
@@ -92,18 +92,16 @@
 							<textarea class="form-control" name="v_fwd" id="v_fwd" <?php if($v_blackhole == 'yes') echo "disabled";?>><?=htmlentities(trim($v_fwd, "'"))?></textarea>
 						</div>
 						<div class="form-check">
-							<input class="form-check-input" type="checkbox" name="v_blackhole" id="v_blackhole" <?php if ($v_blackhole == 'yes') echo 'checked' ?>>
+							<input class="form-check-input js-discard-all-mail" type="checkbox" name="v_blackhole" id="v_blackhole" <?php if ($v_blackhole == 'yes') echo 'checked' ?>>
 							<label for="v_blackhole">
 								<?= _("Discard all mail") ?>
 							</label>
 						</div>
-						<div id="id_fwd_for" style="display:<?php if ($v_blackhole == 'yes') {echo 'none';} else {echo 'block';}?> ;">
-							<div class="form-check">
-								<input class="form-check-input" type="checkbox" name="v_fwd_only" id="v_fwd_for" <?php if ($v_fwd_only == 'yes') echo 'checked' ?>>
-								<label for="v_fwd_for">
-									<?= _("Do not store forwarded mail") ?>
-								</label>
-							</div>
+						<div class="form-check <?php if ($v_blackhole == 'yes') { echo 'u-hidden'; } ?>">
+							<input class="form-check-input" type="checkbox" name="v_fwd_only" id="v_fwd_for" <?php if ($v_fwd_only == 'yes') echo 'checked' ?>>
+							<label for="v_fwd_for">
+								<?= _("Do not store forwarded mail") ?>
+							</label>
 						</div>
 						<div class="u-mt10 u-mb10">
 							<label for="v_rate" class="form-label">
